requirements.txt
streamlit
boto3
openai
secrets
python-dotenv
csv